// src/internal/api/models/PendingEvalPayload.ts
// Auto-generated by scripts/generate-client.ts
// DO NOT EDIT MANUALLY

export interface PendingEvalPayload {
  project_id: string;
  eval_name: string;
  judges: { name: string }[];
  examples: {
    example_id: string;
    created_at: string;
    trace_id: string;
    span_id: string;
  }[];
  is_offline: boolean;
  is_behavior: boolean;
}